+++ public/components/common/hooks/index.ts
@@ -24,4 +24,4 @@ export * from './useApiRequest';
 export * from './use-app-config';
 export * from './useRootScope';
 export * from './use_async_action';
-export { useEsSearch } from './useEsSearch';
+export { useEsSearch } from './use-es-search';
