+++ public/components/common/panels/index.ts
@@ -1,2 +1,2 @@
 export * from './panel_split';
-export { AggTable } from './agg_table';
+export { AggTable } from './agg-table';
